[Journal] Moving the create registered report endpoint to the note
