+++ src/note/views/note_view.py
@@ -7,6 +7,7 @@ from django.shortcuts import get_object_or_404
 from rest_framework import status
 from rest_framework.decorators import action
 from rest_framework.permissions import AllowAny, IsAuthenticated
+from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet
 
@@ -43,7 +44,12 @@ from researchhub_document.related_models.constants.document_type import (
     NOTE,
     REGISTERED_REPORT,
 )
+from researchhub_document.serializers.researchhub_post_serializer import (
+    RegisteredReportDraftSerializer,
+)
+from researchhub_document.services.journal_entry_service import JournalEntryService
 from user.models import Organization, User
+from user.permissions import IsModerator, UserIsEditor
 
 
 class NoteViewSet(ModelViewSet):
@@ -184,6 +190,34 @@ class NoteViewSet(ModelViewSet):
         data = serializer.data
         note.notify_note_created()
         return Response(data, status=200)
+
+    @action(
+        detail=False,
+        methods=["post"],
+        permission_classes=[UserIsEditor | IsModerator],
+        url_path="create_registered_report_draft",
+    )
+    def create_registered_report_draft(self, request: Request) -> Response:
+        """Create an editor- or moderator-owned registered report notebook draft."""
+        serializer = RegisteredReportDraftSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+
+        try:
+            draft = JournalEntryService().create_registered_report_draft(
+                request.user,
+                serializer.validated_data["proposal_id"],
+            )
+        except ValueError as error:
+            return Response(
+                {"error": str(error)},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+        response_data = NoteSerializer(draft.note, context={"request": request}).data
+        response_data["fundraise_id"] = draft.fundraise.id
+        response_data["journey_id"] = draft.journey.id
+        response_data["proposal_id"] = draft.proposal.id
+        return Response(response_data, status=status.HTTP_201_CREATED)
 
     def _create_unified_doc(self, request):
         data = request.data
